docs: Switched to 'useradd', as it is a more cross-distro cmd.
'adduser' is installed on Debian-based systems by default, but is
not installed by default on Fedora-based systems. 'useradd' is
installed by default on Debian-based systems, as well. The
'useradd'-based command provides the same functionality as
'adduser' in this case, though setting up a 'system' user with
no password.

Per bug 886, I've added the '--user-group' flag, which creates a
mediagoblin group, and assigns the mediagoblin user to this
group, too.