docs: Removed note about installing from master. Clarified root/nonroot usage.
There was a note informing the user that they should be aware that we
encourage them to install from the *master* branch, but we no longer
do that. If they follow our deployment docs, they will pull in &
use the *stable* branch.

Also clarified some root / non-root system usage by either specifying
a leading *hash* symbol to indicate a root-based command, or a
dollar sign to indicate a non-root command.