docs: Added handling of user_dev permissions
The user_dev directory gets created when users get established
and they start uploading media. However, the permissions aren't
correct on the directory as it gets created.

As a workaround, we create the directory at this stage, and then
set the proper permissions.

See this mailing list post for more information:
http://lists.mediagoblin.org/pipermail/devel/2015-May/001201.html