Fix Python 2 package dependencies and sub-dependencies [#5595].
This fixes the installation process *right now* on Python 2. The problem is that
packages are dropping Python 2 support all over the place and due to
our (deliberate) loose dependency version specifications and the loose
sub-dependency version specifications, it's one horrible mess. But it works
right now.

Next up I'll be pinning all the currently installed Python 2 packages and all
dependencies.