Refactor debian-python3-sqlite Dockerfile
This is a proof of concept, and if we like it, I can perform the same refactoring on the other Dockerfiles.

This pushes more runtime activities (tests, sqlite initialization, configuration) to runtime instead of baking them into the Docker image.

It also adds files to the .dockerignore to speed up the build.

Most importantly, it uses the local copy of MediaGoblin's source rather than pulling it from an external git repository.